[fix] navbar continers as components + deleted stores
