+++ src/components/SpecComponents.js
@@ -1,7 +1,7 @@
 import Loading from './basics/loading';
 import Modal from './modal/modal';
-import NavbarLogin from './navbar/navbar-login/NavbarLogin.container';
-import NavbarApp from './navbar/navbar-app/NavbarApp.container';
+import NavbarLogin from './navbar/navbar-login/NavbarLogin';
+import NavbarApp from './navbar/navbar-app/NavbarApp';
 import TextInput from './inputs/TextInput';
 import ErrorMessage from './basics/ErrorMessage';
 import Button from './buttons/Button';
