--- src/components/navbar/navbar-login/NavbarLogin.actions.js
@@ -1,4 +1,0 @@
-
-export const SELECTED_ITEM = 'SELECTED_ITEM';
-
-export const selectItem = item => ({ type: SELECTED_ITEM, payload: { item } });
